handwritten-digit-recognition/requirements.txt
tensorflow>=2.10
numpy
pillow
scikit-learn
matplotlib
tk 
